requirements.txt
tkinter
matplotlib
pillow
pandas